dom/media/test/test_bug1120222.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1120222
-->
<head>
  <title>Test for Bug 1120222</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">
var manager = new MediaTestManager;

let count = 0;

function startTest(test, token) {
  manager.started(token);
  var elemType = /^audio/.test(test.type) ? "audio" : "video";
  var v = document.createElement(elemType);
  v.token = token;
  v.src = test.name;

  v.addEventListener("play", (e) => {
    document.title = `Test Title ${count++}`;
    ok(true, "changed title while media was playing")
    removeNodeAndSource(e.target);
    manager.finished(e.target.token);
  });

  document.body.appendChild(v);
  v.play();
}

manager.runTests(gSmallTests, startTest);

</script>
</pre>
</body>
</html>
